lock,store: add support for windows
This was a bit of a wild ride, because Windows, by default, doesn't
allow files to be opened for reading/writing/deletion by multiple
processes, and because there aren't many ways to perform an atomic
rename that overwrites the destination unconditionally, even when
opened by another process.

This special set of circumstances meant that os.OpenFile and os.Rename
could not be used by the portable Store code, requiring our own
implementation. The functions are replaced by openShared and rename
respectively, which default to os.OpenFile and os.Rename on all
platforms except Windows, which provides its own custom implementations.